fix: apply dirty editor suppression to notebooks
Use the same generic open-when-dirty suppression for notebook chat edits as
for text documents, and have NotebookEditorManager honor that override before
opening dirty notebook editors.
